requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
requests
pypdf
openpyxl
openai-agents
cuid

psycopg2-binary>=2.9.9